support systemd

+++ common/package.mk
@@ -1,21 +1,18 @@
 SHELL := /bin/bash
-DOWNLOAD := $(DOWNLOAD)/package
-STATE := $(STATE)/package
-host_path := $(PATH)
-PREFIX ?= rootfs
-prefix := $(PREFIX)
-export PATH := $(TOOLCHAIN)/bin:$(PATH)
-export CPPFLAGS := -I$(STAGE)/$(prefix)/include -L. -fPIC
-export CFLAGS := -I$(STAGE)/$(prefix)/include -L. -fPIC
-export LDFLAGS := -L$(STAGE)/$(prefix)/lib
+DOWNLOAD := $(DOWNLOAD)
+STATE := $(STATE)
+ROOT_PREFIX ?= sysroot
+LOCAL_BUILD ?= 0
 
 .PHONY: all
 -include $(BASE)/../packages/*/*.mk
 
-depends = $(subst $(BASE),./,$(STATE))/dep.$1.$2
+
+depends = $(subst $(BASE),./,$(STATE))/dep.$1.$2.build_$(LOCAL_BUILD)
 
 define depfile =
 	mkdir -p '$(STATE)'
+	$(info "depfile: $(BASE)/$(call depends,$1,$2)")
 	touch '$(BASE)/$(call depends,$1,$2)'
 endef
 
@@ -23,39 +20,40 @@ endef
 # Dynamically declare dependencies between packages
 #############################################
 define declaredeps =
-$(eval .PHONY: $1)
-$(eval $1: $(call depends,$1,package))
-$(eval $(call depends,$1,package) : $(call depends,$1,install) ; $(call packagepkg,$1) )
-$(eval $(call depends,$1,install) : $(call depends,$1,build)   ; $(call installpkg,$1) )
-$(eval $(call depends,$1,build)   : $(call depends,$1,prepare) ; $(call buildpkg,$1)   )
-$(eval $(call depends,$1,prepare) : $(call depends,$1,download); $(call preparepkg,$1) )
-$(eval $(call depends,$1,download):                            ; $(call downloadpkg,$1))
-
-$(eval $(call depends,$1,build)   : $(foreach dep,$($1/DEPENDS),$(call depends,$(dep),install)))
-$(foreach dep,$($1/DEPENDS),$(call declareonce,$(dep)))
+	$(eval .PHONY: $1)
+	$(eval $1: $(call depends,$1,package))
+	$(eval $(call depends,$1,package) : $(call depends,$1,install) ; $(call packagepkg,$1) )
+	$(eval $(call depends,$1,install) : $(call depends,$1,build)   ; $(call installpkg,$1) )
+	$(eval $(call depends,$1,build)   : $(call depends,$1,prepare) ; $(call buildpkg,$1)   )
+	$(eval $(call depends,$1,prepare) : $(call depends,$1,download); $(call preparepkg,$1) )
+	$(eval $(call depends,$1,download):                            ; $(call downloadpkg,$1))
 endef
 
-define declareonce =
+define declare_all =
 	$(if $($1_done),,$(call declaredeps,$1) $(eval $1_done=1))
 endef
 
 .SHELLFLAGS = -e -c
 .ONESHELL:
 
+
 #############################################
 # Macros for specific stages
 #############################################
 define downloadpkg =
+	$(eval $(info "DOWNLOAD_PKG: $(LOCAL_BUILD) URL: $($1/TARBALL)"))
 	mkdir -p '$(DOWNLOAD)'
 	cd '$(DOWNLOAD)'
-	$(info DOWNLOAD:  $(DOWNLOAD))
 	if [ -n '$($1/TARBALL)' ]; then
-		wget '$($1/TARBALL)' -O- > '$(DOWNLOAD)/$(notdir $($1/TARBALL))'
+		if [ ! -f $(notdir $($1/TARBALL)) ]; then
+			wget '$($1/TARBALL)' -O- > '$(DOWNLOAD)/$(notdir $($1/TARBALL))'
+		fi
 	fi
 	$(call depfile,$1,download)
 endef
 
 define preparepkg =
+	$(eval $(info "PREPARE_PKG"))
 	mkdir -p '$(BUILD)/$1'
 	cd '$(BUILD)/$1'
 	if [ -n '$($1/TARBALL)' ]; then
@@ -66,18 +64,21 @@ define preparepkg =
 endef
 
 define buildpkg =
-	mkdir -p '$(STAGE)'
+	$(eval $(info "BUILD_PKG"))
+	mkdir -p '$(HOST)'
 	$(call $1/build)
 	$(call depfile,$1,build)
 endef
 
 define installpkg =
+	$(eval $(info "INSTALL_PKG"))
 	$(call $1/install)
 	$(call depfile,$1,install)
 endef
 
 define packagepkg =
-	mkdir -p '$(OUT)'
+	$(eval $(info "PACKAGE_PKG"))
+	mkdir -p '$(OUTPUT)'
 	$(call $1/package)
 	$(call depfile,$1,package)
 endef
@@ -88,8 +89,8 @@ endef
 all: $(PACKAGES)
 
 # Import dependencies between packages and package stages
-$(foreach pkg,$(PACKAGES),$(call declareonce,$(pkg)))
+$(foreach pkg,$(PACKAGES),$(call declare_all,$(pkg)))
 
 clean:
-	rm -rf '$(DOWNLOAD)' '$(STATE)' '$(STAGE)' '$(OUT)' '$(BUILD)'
+	rm -rf '$(DOWNLOAD)' '$(STATE)' '$(STAGE)' '$(OUTPUT)' '$(BUILD)'
